Extract scribe from collector
